Renamed some properties to match coventions, changed float types to decimal #18

+++ Source/hatfield.envirodata.dataacquisition.esdat/ChemistryFileData.cs
@@ -14,11 +14,11 @@ namespace Hatfield.EnviroData.DataAcquisition.ESDAT
         public string FieldID { get; set; }
         public double Blank1 { get; set; }
         public double SampleDepth { get; set; }
-        public float Blank2 { get; set; }
+        public double Blank2 { get; set; }
         public string MatrixType { get; set; }
         public string SampleType { get; set; }
         public string ParentSample { get; set; }
-        public float Blank3 { get; set; }
+        public double Blank3 { get; set; }
         public string SDG { get; set; }
         public string Lab_Name { get; set; }
         public string LabSampleID { get; set; }
